move resultsToXml to reportBuilder
